Move AuthError to auth feature
