+++ src/features/auth/AuthError.tsx
@@ -1,7 +1,7 @@
 import { signIn } from "next-auth/react";
 import Link from "next/link";
 
-export default function AuthError() {
+export function AuthError() {
   return (
     <div>
       <h2>No estas autorizado a acceder a esta pagina</h2>
